pack_kernel.sh support argument.

+++ pack_kernel.sh
@@ -7,8 +7,12 @@ TMP_IMG=./tmp_img
 
 rm -rf ${SRC_KERNEL}
 
-#cat arch/arm/boot/zImage arch/arm/boot/dts/210-sate210.dtb > ${SRC_KERNEL}
-cat arch/arm/boot/zImage arch/arm/boot/dts/210-tq210v4.dtb > ${SRC_KERNEL}
+if [ ! -f arch/arm/boot/dts/$1 ]; then
+	echo "not found arch/arm/boot/dts/$1"
+	exit -1
+fi
+
+cat arch/arm/boot/zImage arch/arm/boot/dts/$1 > ${SRC_KERNEL}
 
 rm -rf ${DST_KERNEL}
